@@ -36,10 +36,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  google_maps_flutter: ^2.5.3
+  google_maps_flutter: ^2.6.0
   flutter_polyline_points: ^2.0.0
-  http: ^1.2.0
-  logger: ^2.1.0
+  http: ^1.2.1
+  logger: ^2.2.0
   intl: ^0.18.1
   shared_preferences: ^2.2.2
   global_configs: ^1.1.2
@@ -52,21 +52,18 @@ dependencies:
   sliding_up_panel: ^2.0.0+1
   page_transition: ^2.1.0
   flutter_masked_text2: ^0.9.1
-  appmetrica_plugin: ^1.4.0
+  appmetrica_plugin: ^2.0.0
   audioplayers: ^5.2.1
-  firebase_core: ^2.27.2
+  firebase_core: ^2.28.0
   firebase_messaging: ^14.7.21
   uuid: ^4.3.3
   android_play_install_referrer: ^0.3.0
-  package_info_plus: ^4.2.0
+  package_info_plus: ^6.0.0
   google_fonts: ^6.2.1
   vector_math: ^2.1.4
   fluttertoast: ^8.2.4
   permission_handler: ^11.3.1
   platform_device_id_v3: ^1.0.0
-
-
-  
 
 dev_dependencies:
   flutter_test:
